pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "newsapp"
version = "0.1.0"
description = "A fast, terminal-based news application displaying headlines from RSS feeds and web sources"
readme = "docs/README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.0,<4.0.0",
    "textual>=0.20.0,<1.0.0",
    "feedparser>=6.0.0,<7.0.0",
    "beautifulsoup4>=4.11.0,<5.0.0",
    "lxml>=4.9.0,<5.0.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=0.20.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "rich>=13.0.0,<14.0.0",
    "structlog>=23.0.0,<24.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "aiohttp-caching>=0.3.0,<1.0.0",
    "typing-extensions>=4.0.0,<5.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "charset-normalizer>=3.0.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.11.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/newsapp"
Documentation = "https://github.com/yourusername/newsapp/blob/main/docs"
Repository = "https://github.com/yourusername/newsapp"
"Bug Tracker" = "https://github.com/yourusername/newsapp/issues"

[project.scripts]
newsapp = "src.main:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html"
